refact: move create_image_bitmap to ImageBitmap Impl (#37348)
