+++ components/script/dom/window.rs
@@ -152,7 +152,7 @@ use crate::dom::storage::Storage;
 #[cfg(feature = "bluetooth")]
 use crate::dom::testrunner::TestRunner;
 use crate::dom::trustedtypepolicyfactory::TrustedTypePolicyFactory;
-use crate::dom::types::UIEvent;
+use crate::dom::types::{ImageBitmap, UIEvent};
 use crate::dom::webglrenderingcontext::WebGLCommandSender;
 #[cfg(feature = "webgpu")]
 use crate::dom::webgpu::identityhub::IdentityHub;
@@ -1209,9 +1209,16 @@ impl WindowMethods<crate::DomTypeHolder> for Window {
         options: &ImageBitmapOptions,
         can_gc: CanGc,
     ) -> Rc<Promise> {
-        let p = self
-            .as_global_scope()
-            .create_image_bitmap(image, 0, 0, None, None, options, can_gc);
+        let p = ImageBitmap::create_image_bitmap(
+            self.as_global_scope(),
+            image,
+            0,
+            0,
+            None,
+            None,
+            options,
+            can_gc,
+        );
         p
     }
 
@@ -1226,7 +1233,8 @@ impl WindowMethods<crate::DomTypeHolder> for Window {
         options: &ImageBitmapOptions,
         can_gc: CanGc,
     ) -> Rc<Promise> {
-        let p = self.as_global_scope().create_image_bitmap(
+        let p = ImageBitmap::create_image_bitmap(
+            self.as_global_scope(),
             image,
             sx,
             sy,
